+++ packages/android/jnilib/src/main/java/com/example/jnilib/model/FoundPrimes.kt
@@ -6,7 +6,7 @@ package com.example.jnilib.model
 class FoundPrimes private constructor() {
 
   val primeCount: Int get() = nativePrimeCount()
-  val upTo: Int get() = nativeUpTo()
+  val upTo: Long get() = nativeUpTo()
   val foundPrimes: List<Long> get() = nativeFoundPrimes().toList()
 
   @Suppress("unused") // used from the JNI side
@@ -16,5 +16,5 @@ class FoundPrimes private constructor() {
 
   private external fun nativePrimeCount(): Int
   private external fun nativeFoundPrimes(): LongArray
-  private external fun nativeUpTo(): Int
+  private external fun nativeUpTo(): Long
 }
